Basic Weather App
This is a simple web application that queries weather conditions and displays the weather forecast for a given city. It was built using JavaScript and CSS, and uses the OpenWeatherMap API to retrieve weather information. Simply enter the name of a city in the search box and click "Search" to view the results.